Update namespaces, use file-scoped namespaces, reflect filenames in classnames
